+++ lib/angleAndPosition/angleAndPosition.h
@@ -79,9 +79,9 @@ class Angle {
     /**
      * @brief Retuns the angles equivelent value in radians.
      *
-     * @return (int16_t) The value of the angle in radians.
+     * @return (double) The value of the angle in radians.
      */
-    int16_t toRadians();
+    double toRadians();
 
     /**
      * @brief Returns the index of the segments at the current angle.
